requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy==2.0.36
mysql-connector-python==9.1.0
python-dotenv==1.0.1
requests==2.32.3
jinja2==3.1.4